+++ lib/sobranie_page.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+require 'scraped'
+
+class SobraniePage < Scraped::HTML
+  decorator Scraped::Response::Decorator::AbsoluteUrls
+end
